Add a `forCache` package with all build dependencies (#34)
We want to ensure that the binary cache has all build dependencies,
so that users do not have to rebuild Torch, gcc, etc. This change
creates a `forCache` package that is a linkFarm with all supported
Torch versions as well as the modified stdenvs that are used for
backwards compat with old Linux versions.