Add Torch 2.6, drop Torch 2.4 (#56)
This requires some changes across the board because Torch 2.6 has a
different set of supported CUDA versions --- it adds CUDA 12.6, dropping
CUDA 12.1.

We were also leaking the default stdenv into the build environment. This
was ok for older nixpkgs versions, because the default gcc version was 13,
which CUDA is compatible with. However, nixpkgs switched to gcc 14,
which is not supported by CUDA. This change rectifies this leakage.

Finally, also make sure that the `torch.{cxxdev,dev}` outputs gets cached.