feat: update docker for remote build and push (#115)
This PR enables the container to run the nix build command as a non root
user. This functionality is required to allow the builder to run on
hfjobs; which sets the user is 1000.

deploy a build job via

```bash
hfjobs run \
    drbh/kernel-builder-dev:job-runner-0 \
    /home/nixuser/bin/cli.sh \
    fetch \
    https://huggingface.co/kernels-community/activation.git
```

notes
- `drbh/kernel-builder-dev:job-runner-0` is an image of the attached
Dockerfile
- `/home/nixuser/bin/cli.sh` is the entrypoint that needs to be
specified on `hfjobs`
- `fetch` is the docker command
- `https://huggingface.co/kernels-community/activation.git` is the
kernel repo to build

currently this feature does not push the built files to the hub but that
can/should be added in a followup PR when a suitable upload strategy is
defined.