Pass default versions set `torchVersions` (#193)
`genFlakeOutputs` accepts a `torchVersions` attr to define custom
versions. However, up till this change it took a list. This made it
impossible to add the existing set of versions. With this change
`torchVersions` must be a one-arg function. It will be passed the
default version set, making it possible to extend the versions:

```nix
kernel-builder.lib.genFlakeOutputs {
  # ...

  torchVersions = defaultVersions: defaultVersions ++ [
    # Custom versions
  ];
};
```