Fix handling of the 9.0a and 12.0a capabilities (#202)
The code that computes the intersection between a kernel's capabilities
and the supported capabilities failed because e.g. the base capability
of 9.0a (9.0) did not match 9.0+PTX.

Instead we should just leave the supported capabilities at regular
version numbers. Instead, make a list of default kernel capabilities and
use that and intersect with the supported capabilities if no
capabilities are specified for a kernel.