Select default shells from available variants (#247)
Before this change, `nix develop` and `nix develop .#test` would use the
`torch28-metal-aarch64-darwin` or `torch28-cxx11-cu126-${system}` build
variants on macOS or Linux respectively. This had two downsides: (1) it
resulted in an error for ROCm/XPU-only kernels, requiring to manually
specify the variant; (2) we had to maintain the variants (e.g. bump up
the Torch version when 2.8 is not supported anymore).

This change chooses the shell depending on the variants that are
available for a given kernel on a given system. We use the following
ordering to sort the available variants:

- Bundle variants before non-bundle variants.
- CUDA variants before other frameworks.
- Newer Torch versions before older versions.
- Older frameworks before newer (best system compatibility).

Then we choose the first variant in this ordering. This should select
the best variant in most cases.