Restructure extension to cache extra ROCm and XPU dependencies (#257)
Before this change the `torch-extension` directory was just a derivation
that built an extension for a given configuration + build set. The
downside of this approach was that we could not easily get things like

- The standard environment to be used by dev shells.
- Overrides for caching. For instance, we override ROCm's `clr` and
XPU's `oneapi-torch-dev` and `onednn-xpu` to use stdenv with an old
glibc.

This change modifies `torch-extension` so that we can instantiate it
given a build set. A set is returned that can give access to the things
mentioned above, as well as mkExtension and mkNoArchExtension functions.
We assign this set to the corresponding build set, so that it is
directly accessible from the build set.

Use this change to add ROCm and XPU overrides to the cache.