Introducing a way to builder kernels for Windows platforms (#250)
This PR attempts to replicate what is being done with NixOS on Windows
platform through a PowerShell scripts `kbuild.ps1`.

It leverages `build2cmake` as much as possible and attempts to move some
parts currently handle in Nix directly in CMake to avoid dupplication as
much as possible (when feasible).

It also introduce two CMake functions to automatically generate the
layout required by `kernels::get_local_kernel` and hub based build, so
it easy to debug and publish.