fix(windows): always define _WIN32 preprocessor macro to prevent PyTorch compiling unsupported code (#275)
PyTorch (especially dynamo) relies on the preprocessor macro `_WIN32` to
guard compilation of code sections in some places when targetting
Windows.

This macro is defined by MSVC defacto but, NVCC and others are not
setting it by default which can lead to compilation errors.

This PR ensures the macro is defined when targetting Windows platform so
we can safely rely its availability for the preprocessor.