Flatten build variants to `build/<variant>` (#293)
Prior to this change, kernels were stored in
`build/<variant>/<extname>`. However, this was fragile because the
extension name had to correspond to the repository name. This change
flattens kernels to be stored inside `build/<variant>`.

For compatibility with older versions of kernels, we add a module
`build/<variant>/<extname>` that loads `build/<variant>`, this
compatibility module will removed when the `kernels` update has been
around for a while.