Copy over Torch from hf-nix and fix the AArch64 build (#304)
We have decided to make kernel-builder standalone to make it easier to
maintain Torch versions as part of the same repo. This copies over Torch
from hf-nix. Other bits are still needed, but this was needed to fix
#301 within the repo.

Torch vendors libgfortran, which has zlib as a dependency. We might want
to use system libgfortran in the future instead, but it requires more
investigation/testing.